lib/scanner: rewrite tag directive scan, return Ref over Token
This commit takes the first steps towards the final API of Scanner,
wherein it returns Result<Ref>s over Result<Token>s. This change allows
the struct to access a scratch space which it can and will use when
borrowing from the underlying data is impossible, such as when
encountering escape sequences (which must be unescaped), line joining,
or data (type) transformation.

Regardless, these changes were required to correctly handle escape
sequences in tag directives.

Note that the test suite for lib/scanner is broken as of this commit, it
will be fixed in the next.