lib/scanner: further fixes to zero indented sequence handling
While the previous commit did add support for _adding_ zero indented
sequences to the token stream, it unfortunately relied on the indent
stack flush that happens once reaching end of stream to push the stored
BlockEnd tokens.

This commit adds better support for removing zero indented sequences
from the stack once finished.

The heuristic used here is:

A zero_indented BlockSequence starts when:

- The top stored indent is for a BlockMapping
- A BlockEntry occupies the same indentation level

And terminates when:

- The top indent stored is a BlockSequence & is tagged as zero indented
- A BlockEntry _does not_ occupy the same indentation level