scalar/plain: add scan_plain_scalar
the primary driver for scanning plain YAML scalars. This implementation
tries to fit as closely as possible to the YAML spec, particularly in
its handling of (the lack of) spacing requirements inside flow contexts,
comment detection and special casing of - ? : as first character in flow
contexts.

Two things that are notably missing:

1. Proper tab '\t' handling in block context indentation
2. A sane maximum whitespace limit && better handling of whitespace
   storage. Rather than storing every whitespace given, I could instead
   count the whitespace separated by line breaks, and then add it back
   later, such that the maximum described above would apply to total
   line breaks, with the intervening whitespace stored as a u64/usize