lib: fix visibility on Queue, Scanner, TokenEntry
Each of these will likely appear in the parts of the public API, even if
they aren't directly used.

Its likely these will be "public but unreachable" -- e.g a public type
in a private module.

This will likely be revisited on the way to a stable 1.0 library
version, but works for now.