build/justfile: add recipes for just
Just is an improvement on Make/Makefiles for people that wish to use it
as a recipe runner rather than a build system.

We are introducing a dependency on it, as it gives us easy ways to run
arbitrarily complex commands in a composable fashion.

In particular, these recipes allow us 5 major gains:

1. Setting our repo's environment consistently (fresh-system, install-bins)
2. Run comprehensive checks on the local tree's code (lint, audit, fmt,
   test, test-all, lint-docs)
3. Utility commands for inspecting dependencies (deps, rdeps, udeps),
   documentation (docs) or git hygiene (git-branch-prune)
4. Provides extra environment vairables that can be used to control
   things like features and profile used when building, which are
   missing from Cargo and/or rustc
5. Provides a uniform interface for repo tasks, abstracting over many
   different underlying tools